Working on Automated Test Generator
The automated test generator is now generating a test case which should
result in a false negative.  Specifically, RuntimeValidTests.test_100()
is the test case in question (which is only temporarily in
RuntimeValidTest) and this illustrates that the machine gets stuck with
a dangling reference exception.

However, the AutomatedTestGenerator does not see this dangling reference
exception for reasons unknown.  It is almost certainly because of two
specific kludges in the generation of blocks.  Specifically, that the
lifetimes given for blocks are garbage.

One potentially easy solution here is to actually remove the
association between a block and its lifetime.  Instead, we can create
lifetimes as necessary within the abstract semantics.  Whilst in
principle that should work, it is a problem I think for small step
semantics where we need to remember within the syntax the lifetime.

The next step is to create a lifetime class which has a "within"
relation.