Add FuzzTestExperiment
This adds an experiment comparing the calculus against the Rust borrow
checker.  At this time, there are a number of inconsistencies thrown up
which need further examination.  For example, sampling from the domain
of all programs produces those with multiple declarations for the same
variable.  These are permitted in Rust, though not by my calculus.
However, sampling from the domain of programs where every variable is
defined before being used results avoids this, though perhaps is not
ideal.