Contemplating options for resolving issues.
At this stage, the main issue is that explicit copy operations do not cause
errors in Rust when they are for unused variables.  That is, the Rust Compiler
simply converts them into move operations.  Unfortunately, I need some way to do
this myself.  Potentially, I could actually have some kind of "undo" which
converts copies into moves when the variables in question are unused.