Add improved inconsistency analysis
This adds some improved analysis of inconsistent results between rustc and FR.
The analysis appears to completely categorise all forms of inconsistency, which
is very nice!  The only downside is that it has identified a previously unknown
limitation with FR regarding the rule T-BoxDeref.  This rule can, in fact, be
relaxed from requiring copyable types to operating on box types under certain
circumstances.