Hopefully final update?
Am still having problems on the grid for some reason.  Using sha-256 hashes
didn't really help.  Instead, am trying to use the local temporary directory on
each node.  Running on a single machine is fine (even with multiple cores), but
running multiple machines together at the same time is causing problems.  Maybe
something to do with NFS.  Unsure, though it was definitely working originally.

Have also noticed one problem related to line information on FR syntax errors.
This appears to be working correctly on my local machine, but not on the grid.
Note that this is one of the changes from the original grid run (which
completed).  But, I cannot for the life of me understand how/why this would be
causing a problem or why it works differently on the grid :|