Eliminate distinction between Stmt / Expr
This puts through a fairly large and radical move which eliminates the
distinction between expressions and statements.  This actually reflects the Rust
language more accurately.  However, it does indeed require wholesale changes to
the FR system.

This also eliminates the BigStepOperationalSemantics.  Hopefully this is a good
descision, since that was more efficient for model checking.  However, the
small-step semantics better reflects the formalism used in the paper.