Add ControlFlow Extension
This adds syntax, semantics and typing rules for if-else statements.  A key
challenge here was the mechanism for merging types at control-flow meet points.
This supports borrows of the form &(x,y) or &mut (x,y) where x and y are
compatible variables.  To make this work, have eliminated strong updates from
the calculus as these do not appear to be supported by Rust.  Have also formally
introduced a "void" type to represent the return of a statement.

This puts in place several fixes for control-flow execution.  This includes
updating the concept of a borrow to refer to multiple locations.  Furthermore,
the concept of a weak update is applied to all indirect assignments.  This
currently models Rust, though could be extended in the future.