Support multi-borrows
This supports borrows of the form &(x,y) or &mut (x,y) where x and y are
compatible variables.  To make this work, have eliminated strong updates from
the calculus as these do not appear to be supported by Rust.  Have also formally
introduced a "void" type to represent the return of a statement.

Still need to check strong updates for boxes.  Likewise, the operational
semantics is failing for control-flow statements.