Add support for pairs.
This adds the initial support for pairs, as originally outlined in the paper.
Specifically, this includes pair constructors and access operators along with
a corresponding pair type.  The type rule outlined in the pair appears to work
(though more extensive testing is required here).

At this stage, there is still a problem related to type compatibility which
needs to be moved out of BorrowChecker and into Type.
Furthermore, we need support for partial borrows.  This will require
changing the nature of a borrow from a single name to a path.