Fixes for Pairs
More fixes for pairs.  At this stage, things are starting to work.  The concept
of a slice is making sense.

There were some bugs related to pair reduction --- fixed now.  Have also added
test cases which illustrate the problem related to the lack of first-class
effects.  In particular, we need to apply effects to part of a type (e.g. to
indicate that one field has been moved whilst the others have not).  In doing
this, we need to retain the overall shape for subsequent assignments.