Support Shadow Effect
This pushes through support for effects which is essentially working now.  A
bunch of interesting questions are raised, but we'll see.  At this stage, still
need to fix up some issues related to control-flow.  Also, have an outstanding
problem cannot fix related to strong updates and boxes.

This puts in place some fixes for borrowing paths.  The system is getting more
complicate by the second and it's going to be interesting trying to simplify it
as much as possible.