Remove copy syntax
This removes the "copy" syntax which was previously used to distinguish moves
and copies.  Instead, the abstract machine now always copies values, including
references.  This can result in dangling references in the machine.  Likewise,
in principle, we could have multiple references to the same location marked as
mutable references if the typing rules were incorrect somehow.

Finished Pairs / Tuples

This marks pairs / tuples as essentially finished.