Refactor paths
This represents a substantial refactoring of the calculus to introduce a general
concept of a "path", rather then the explicit syntax previously used.  This
means, for example, that assignment and indirect assignment are merged into a
single concept of assignment.  The latter includes a path component.  So, for
example, "x = e" is simply an assignment with an empty path;  similarly, "*x =
e" is an assignment with a path containing a single deref.

There are two reasons I've made this change.  Firstly, it helps to reduce the
core calculus and make it much more easily extensible (e.g. for tuples or
structures, etc).  Secondly, it paves the way to support reborrowing which, in
essence, becomes a fairly trivial extension (or indeed is perhaps built in from
the start)