Improved memory deallocation
This puts in place an improved notion of memory deallocation.  At this stage,
the underlying semantics destructive reads for owned references, and any
compounds which contain owned references.  In principle, it might make sense to
reintroduce the concept of explit move / copy.  Otherwise, without this, the
semantics handles a lot of the complicating cases surrounding memory problems
automatically.