Add Explicit Copy Syntax
After everything, I've decided to put explicit copy syntax back.  This is
because it allows the machine to get in various broken states which, without
explicit syntax, it can't.  For example, we can easily end up with two owning
points to the same box with explicit copy syntax.  But, we cannot do that
without it.

At this stage, all syntax has been updated and it remains to go through all the
test cases making copy variants, and then fix the borrow checker to actually
check that a copy operation allowed.