Support explicit copy/move in BorrowChecker
This puts in place support for explicit copy/move syntax into the borrow
checker.  This required some interesting changes for ControlFlow.  Specifically,
we need a notion of "temporary" to capture the situation where values are only
temporarily in the condition only.

There was also a bug related to blocks returning values which has now been
fixed.