Exploit Walker.advance()
JModelGen 0.4.2 provides a new method Walker.advance() which can improve (quite
significantly) the time it takes to progress through an irregular space upto the
start of a given batch.  This makes certain operation quicker (e.g. counting the
size of an irregular domain) and also should help with running on the grid.

This also fixes an outstanding issue which I believe is related to rounding of
batches and attempts to iterate past the end of a space.