Add OperationalSemantics.execute(); tidy up.
This adds an outermost method for executing terms until completion.  In the
context of small-step semantics this is essentially a loop which repeats until
either a value is returned or the machine is stuck (as indicated by an
exception).

This also tidies up things a little by moving as much stuff as possible to be
protected, rather than public.