Add tests and fix for T-Block bug
There was a bug in the T-Block rule related to values being passed out of
blocks.  Specifically, they weren't being checked for lifetime inclusion.  This
was identified during reproving the system, which highlights the benefit of it. I
first added some test cases to verify this, and the added the necessary check to
the rule.  All seems well now.