Support function invocations
This puts through support for function invocations.  Have uncovered and
fixed a small bug related to the heap invariants not correctly
accounting for temporary values.  This was only an issue for checking
heap invariants, not for the actual semantics / typing itself.

At this stage, functions are implemented all the way upto side-effects.
However, subtyping of mutable borrows is not handled correctly as it is
co-variant.